Heat Recipes: Do not alter 'hidden' or 'enabled' properties
The recipe we are copying should already have these properties set
correctly, so do not alter them.

Note: The properties 'hidden' and 'enabled' can be either an entry in
the recipe itself or entries in the recipe.normal and recipe.expensive
entries.

Not altering these properties saves us some time figuring out the
correct place to modify.